+++ TapHeadingAndroid/Assets/GooglePlayPlugins/com.google.android.appbundle/Editor/Scripts/AndroidAppBundle.cs
@@ -12,18 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// In the Unity 2017 series the EditorUserBuildSettings.buildAppBundle field was introduced in 2017.4.17.
-// It might seem preferable to modify buildAppBundle using reflection, but the field is extern.
-// Instead check for quite a few versions in the 2017.4.17+ series.
-// NOTE: this supports up to UNITY_2017_4_50 and will have to be extended if additional versions are released.
-
-#if UNITY_2018_3_OR_NEWER || UNITY_2017_4_17 || UNITY_2017_4_18 || UNITY_2017_4_19 || UNITY_2017_4_20 || UNITY_2017_4_21 || UNITY_2017_4_22 || UNITY_2017_4_23 || UNITY_2017_4_24 || UNITY_2017_4_25 || UNITY_2017_4_26 || UNITY_2017_4_27 || UNITY_2017_4_28 || UNITY_2017_4_29 || UNITY_2017_4_30 || UNITY_2017_4_31 || UNITY_2017_4_32 || UNITY_2017_4_33 || UNITY_2017_4_34 || UNITY_2017_4_35 || UNITY_2017_4_36 || UNITY_2017_4_37 || UNITY_2017_4_38 || UNITY_2017_4_39 || UNITY_2017_4_40 || UNITY_2017_4_41 || UNITY_2017_4_42 || UNITY_2017_4_43 || UNITY_2017_4_44 || UNITY_2017_4_45 || UNITY_2017_4_46 || UNITY_2017_4_47 || UNITY_2017_4_48 || UNITY_2017_4_49 || UNITY_2017_4_50
-#define GOOGLE_ANDROID_APP_BUNDLE_HAS_NATIVE_SUPPORT
-using UnityEditor;
-#endif
 using System;
 using System.Text.RegularExpressions;
 using Google.Android.AppBundle.Editor.Internal.Utils;
+using UnityEditor;
 
 namespace Google.Android.AppBundle.Editor
 {
@@ -34,9 +26,14 @@ namespace Google.Android.AppBundle.Editor
     public static class AndroidAppBundle
     {
         /// <summary>
-        /// The base module contains the Unity game engine.
+        /// Reserved name for the base module that contains the Unity game engine.
         /// </summary>
         public const string BaseModuleName = "base";
+
+        /// <summary>
+        /// Reserved name for the module that separates the base module's assets into an install-time asset pack.
+        /// </summary>
+        public const string BaseAssetsModuleName = "base_assets";
 
         /// <summary>
         /// Regex used to determine whether a module name is valid.
@@ -46,59 +43,61 @@ namespace Google.Android.AppBundle.Editor
 
         /// <summary>
         /// Returns true if the specified name is a valid Android App Bundle module name, false otherwise.
-        /// The name "base" is reserved for the base module, so also return false in that case.
+        /// Certain names like "base" are reserved, so also return false in those cases.
         /// </summary>
         public static bool IsValidModuleName(string name)
         {
             // TODO(b/131241163): enforce a name length limit if we make it much smaller than 65535.
             return name != null
                    && NameRegex.IsMatch(name)
-                   && !name.Equals(BaseModuleName, StringComparison.OrdinalIgnoreCase);
+                   && CheckReservedName(name, BaseModuleName)
+                   && CheckReservedName(name, BaseAssetsModuleName);
         }
 
         /// <summary>
-        /// Returns true if this version of the Unity Editor has native support for building an Android App Bundle,
-        /// and false otherwise.
+        /// Always returns true. Previously indicated if this version of the Unity Editor has native support for
+        /// building an Android App Bundle.
         /// </summary>
+        // TODO(b/189958664): Needed for 1.x API compatibility. Should be removed with 2.x.
+        [Obsolete("This is always true")]
         public static bool HasNativeBuildSupport()
         {
-#if GOOGLE_ANDROID_APP_BUNDLE_HAS_NATIVE_SUPPORT
             return true;
-#else
-            return false;
-#endif
         }
 
         /// <summary>
         /// Returns EditorUserBuildSettings.buildAppBundle if it is defined and false otherwise.
         /// </summary>
+        // TODO(b/189958664): Needed for 1.x API compatibility. Should be removed with 2.x.
+        [Obsolete("Use EditorUserBuildSettings.buildAppBundle directly instead")]
         public static bool IsNativeBuildEnabled()
         {
-#if GOOGLE_ANDROID_APP_BUNDLE_HAS_NATIVE_SUPPORT
             return EditorUserBuildSettings.buildAppBundle;
-#else
-            return false;
-#endif
         }
 
         /// <summary>
         /// Enable the EditorUserBuildSettings.buildAppBundle field if it is defined.
         /// </summary>
+        // TODO(b/189958664): Needed for 1.x API compatibility. Should be removed with 2.x.
+        [Obsolete("Use EditorUserBuildSettings.buildAppBundle directly instead")]
         public static void EnableNativeBuild()
         {
-#if GOOGLE_ANDROID_APP_BUNDLE_HAS_NATIVE_SUPPORT
             EditorUserBuildSettings.buildAppBundle = true;
-#endif
         }
 
         /// <summary>
         /// Disable the EditorUserBuildSettings.buildAppBundle field if it is defined.
         /// </summary>
+        // TODO(b/189958664): Needed for 1.x API compatibility. Should be removed with 2.x.
+        [Obsolete("Use EditorUserBuildSettings.buildAppBundle directly instead")]
         public static void DisableNativeBuild()
         {
-#if GOOGLE_ANDROID_APP_BUNDLE_HAS_NATIVE_SUPPORT
             EditorUserBuildSettings.buildAppBundle = false;
-#endif
+        }
+
+        private static bool CheckReservedName(string name, string reserved)
+        {
+            return !name.Equals(reserved, StringComparison.OrdinalIgnoreCase);
         }
     }
 }