+++ TapHeadingAndroid/Assets/GooglePlayPlugins/com.google.android.appbundle/Editor/Scripts/Internal/BuildTools/AndroidAssetPackagingTool.cs
@@ -15,7 +15,6 @@
 using System;
 using System.IO;
 using Google.Android.AppBundle.Editor.Internal.PlayServices;
-using Google.Android.AppBundle.Editor.Internal.Utils;
 
 namespace Google.Android.AppBundle.Editor.Internal.BuildTools
 {
@@ -25,17 +24,11 @@ namespace Google.Android.AppBundle.Editor.Internal.BuildTools
     public class AndroidAssetPackagingTool : IBuildTool
     {
         /// <summary>
-        /// Minimum version of Android SDK Build-Tools where aapt2 supports the "convert" command.
+        /// Minimum version of Android SDK Build-Tools that supports "aapt2 link --proto-format".
         /// </summary>
         private const string BuildToolsMinimumVersion = "28.0.0";
 
-        /// <summary>
-        /// Latest version of Android SDK Build-Tools known to be compatible with this plugin.
-        /// </summary>
-        private const string BuildToolsLatestVersion = "28.0.3";
-
         private const string BuildToolsDisplayName = "Android SDK Build-Tools";
-        private const string BuildToolsPackageName = "build-tools;" + BuildToolsLatestVersion;
 
         private readonly AndroidBuildTools _androidBuildTools;
         private readonly AndroidSdkPlatform _androidSdkPlatform;
@@ -68,27 +61,11 @@ namespace Google.Android.AppBundle.Editor.Internal.BuildTools
                 return true;
             }
 
-            var message = string.Format(
-                "App Bundle creation requires {0} version {1} or later.\n\nClick \"OK\" to install {0} version {2}.",
-                BuildToolsDisplayName, BuildToolsMinimumVersion, BuildToolsLatestVersion);
-            if (buildToolLogger.DisplayActionableErrorDialog(message))
-            {
-                AndroidSdkPackageInstaller.InstallPackage(BuildToolsPackageName, BuildToolsDisplayName);
-            }
-
+            var message =
+                string.Format(
+                    "This build requires {0} version {1} or later.", BuildToolsDisplayName, BuildToolsMinimumVersion);
+            buildToolLogger.DisplayErrorDialog(message);
             return false;
-        }
-
-        /// <summary>
-        /// Given the specified APK, produces a new APK where resources.arsc is converted to proto format.
-        /// </summary>
-        /// <returns>An error message if there was a problem running aapt2, or null if successful.</returns>
-        public virtual string Convert(string inputPath, string outputPath)
-        {
-            return Run(
-                "convert {0} --output-format proto -o {1}",
-                CommandLine.QuotePath(inputPath),
-                CommandLine.QuotePath(outputPath));
         }
 
         /// <summary>