+++ TapHeadingAndroid/Assets/GooglePlayPlugins/com.google.play.common/CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog for com.google.play.common
 
+## [1.8.1] - 2022-09-08
+- Updated plugin's data collection procedure. For more information and the opt-out
+  process, please refer to the [data collection](https://github.com/google/play-unity-plugins#data-collection)
+  section in README.
+
+## [1.8.0] - 2022-07-06
+### New Features
+- Incremented version number to match other packages
+
+## [1.7.0] - 2022-02-15
+### New Features
+- Incremented version number to match other packages
+
+## [1.6.1] - 2022-01-24
+### New Features
+ - Incremented version number to match other packages
+
+## [1.6.0] - 2021-11-15
+### New Features
+ - Incremented version number to match other packages
+
+## [1.5.0] - 2021-06-14
+### Other
+ - Removed ability to compile plugin with Unity 5.6, 2017.1, 2017.2, 2017.3, 2018.1, and 2018.2
+
 ## [1.4.0] - 2021-03-08
 ### Bug Fixes
  - Fixed a build error in 2020.2 related to isHttpError/isNetworkError