+++ TapHeadingAndroid/Assets/GooglePlayPlugins/com.google.play.core/CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog for com.google.play.core
 
+## [1.8.1] - 2022-09-08
+- Updated plugin's data collection procedure. For more information and the opt-out
+  process, please refer to the [data collection](https://github.com/google/play-unity-plugins#data-collection)
+  section in README.
+
+## [1.8.0] - 2022-07-06
+### New Features
+- Migrated to [Core-Common 2.0.0](https://developer.android.com/reference/com/google/android/play/core/release-notes-common#2-0-0).
+## [1.7.0] - 2022-02-15
+### New Features
+- Incremented version number to match other packages
+
+## [1.6.1] - 2022-1-20
+### New Features
+ - Updated Java Play Core dependency from 1.10.2 to 1.10.3
+
+## [1.6.0] - 2021-11-15
+### New Features
+ - Updated Java Play Core dependency from 1.10.0 to 1.10.2
+### Other
+ - Removed playcore.aar file and switched to EDM4U for dependency management
+
+## [1.5.0] - 2021-06-14
+### Other
+ - Removed ability to compile plugin with Unity 5.6, 2017.1, 2017.2, 2017.3, 2018.1, and 2018.2
+
 ## [1.4.0] - 2021-03-08
 ### New Features
  - Updated playcore.aar from 1.8.2 to 1.10.0